requirements.txt
flask
apscheduler
requests
beautifulsoup4
caldav
python-dateutil
pytz
icalendar
